dom/tests/mochitest/general/test_bug1208217.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1208217
-->
<head>
  <title>Test for Bug 1208217</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1208217">Mozilla Bug 1208217</a>
<script type="application/javascript">

  add_task(async function() {
    let pasteCount = 0;
    document.addEventListener('paste', function() {
      pasteCount++;
    });

    is(pasteCount, 0);

    synthesizeKey("V", {accelKey: true});

    is(pasteCount, 1);
  });

</script>
</body>
</html>
